pruba delete

+++ src/main/java/Controller/Actions/ProductosAction.java
@@ -3,38 +3,21 @@ package Controller.Actions;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import Model.*;
-
 import java.util.ArrayList;
 
 public class ProductosAction implements IAction {
     @Override
-    //ACTION=PRODUCTOS.FIND_ALL+Nombre="Teriyaki"+Precio="10.7"
     public String execute(HttpServletRequest request, HttpServletResponse response, String action) {
-        String strReturn ="";
-        switch (action)
-        {
+        String strReturn = "";
+        switch (action) {
             case "FILTER":
-                //strReturn = findByFilter(request, response);
+                // strReturn = findByFilter(request, response);
                 break;
             case "FIND_ALL":
-                Productos prod = new Productos();
-                //prod.setNombre(null);
-                //prod.setPrecioVenta(10.7);
                 strReturn = findAll();
-                //strReturn = findAll(prod);
                 break;
             case "DELETE":
-                String idParam = request.getParameter("ID_PRODUCTO");
-                if (idParam != null) {
-                    try {
-                        int id = Integer.parseInt(idParam);
-                        strReturn = delete(id);
-                    } catch (NumberFormatException e) {
-                        strReturn = "ERROR. Invalid ID format";
-                    }
-                } else {
-                    strReturn = "ERROR. Missing ID parameter";
-                }
+                strReturn = delete(request);
                 break;
             case "ADD":
                 strReturn = add(request);
@@ -45,39 +28,37 @@ public class ProductosAction implements IAction {
         return strReturn;
     }
 
-    private String findAll(/*Productos prod*/) {
-
+    private String findAll() {
         ProductosDao productosDao = new ProductosDao();
-        //ArrayList<Productos> productos = productosDao.findAll(prod);
         ArrayList<Productos> productos = productosDao.findAll(null);
         return Productos.toArrayJSon(productos);
     }
 
-    private String delete(int id) {
-        ProductosDao productosDao = new ProductosDao();
-        int result = productosDao.delete(id);
-        if (result > 0) {
-            return "Producto eliminado con éxito";
+    private String delete(HttpServletRequest request) {
+        String idParam = request.getParameter("ID_PRODUCTO");
+        if (idParam != null) {
+            try {
+                int id = Integer.parseInt(idParam);
+                ProductosDao productosDao = new ProductosDao();
+                int result = productosDao.delete(id);
+                if (result > 0) {
+                    return "Producto eliminado con éxito";
+                } else {
+                    return "ERROR. No se pudo eliminar el producto";
+                }
+            } catch (NumberFormatException e) {
+                return "ERROR. Invalid ID format";
+            }
         } else {
-            return "ERROR. No se pudo eliminar el producto";
+            return "ERROR. Missing ID parameter";
         }
     }
-
-    /*private String findByFilter(HttpServletRequest request, HttpServletResponse response) {
-        ProductosDao productosDao = new ProductosDao();
-        String ID_CATEGORIA_PRD = request.getParameter("FILTRO");
-        Productos productos = new Productos();
-        productos.setIdCategoria(ID_CATEGORIA_PRD );
-        ArrayList<Productos> producto = productosDao.findAllByCategory("");
-        return Productos.toArrayJSon(producto);
-    }*/
 
     private String add(HttpServletRequest request) {
         ProductosDao productosDao = new ProductosDao();
         Productos producto = new Productos();
 
         try {
-            // Obtener los parámetros del request y establecerlos en el objeto Productos
             producto.setIdProducto(request.getParameter("ID_PRODUCTO"));
             producto.setNombre(request.getParameter("PRD_NOMBRE"));
             producto.setPrecioVenta(Double.parseDouble(request.getParameter("PRD_PRECIO_VENTA")));
@@ -99,4 +80,3 @@ public class ProductosAction implements IAction {
         }
     }
 }
-
